[Fix #42] Prevent duplicate donors
This change fixes the issue where creating a donation from the quick
donation modal would create a duplicate donor for the donation
instead of it being attributed to the existing donor.

It also improves the seeds, so that there's now one donor with no
donations, one with one donation, and one with several.