Fix flickering test in `donor_spec.rb`
We unwittingly created a flickering test when we used the id `1` in
our assertion. This would cause the test to sometimes pass, and
sometimes fail, depending on the order in which the tests are run.

This change fixes that by using the id `-1`, which no record can
ever have in the database.